refactor: rearrange & move tabs page to main page group
