+++ src/app/(main)/(tabs)/_layout.tsx
@@ -1,0 +1,8 @@
+// components
+import AppTabs from '@/components/app-tabs';
+
+// ----------------------------------------------------------------------
+
+export default function TabLayout() {
+  return <AppTabs />;
+}
